example_mnist/requirements.txt
datasets==2.12.0
numpy==1.26.2
PyYAML==6.0.1
torch==2.1.0
torchinfo==1.8.0
torchvision==0.16.0
tqdm==4.65.0
wandb==0.15.12